RenderTarget moved to its own file(s)
